# Kyber: AVX2 optimised forward NTT