--- a/pke/kyber/internal/common/asm/go.mod
+++ b/pke/kyber/internal/common/asm/go.mod
@@ -1,0 +1,10 @@
+module github.com/cloudflare/circl/pke/kyber/internal/common/asm
+
+go 1.12
+
+require (
+	github.com/cloudflare/circl v0.0.0
+	github.com/mmcloughlin/avo v0.0.0-20200523190732-4439b6b2c061
+)
+
+replace github.com/cloudflare/circl => ../../../../../