--- a/pke/kyber/internal/common/params.go
+++ b/pke/kyber/internal/common/params.go
@@ -1,15 +1,19 @@
 package common
 
+import (
+	"github.com/cloudflare/circl/pke/kyber/internal/common/params"
+)
+
 const (
 	// Q is the parameter q ≡ 3329 = 2¹¹ + 2¹⁰ + 2⁸ + 1.
-	Q int16 = 3329
+	Q = params.Q
 
 	// N is the parameter N: the length of the polynomials
-	N int = 256
+	N = params.N
 
 	// PolySize is the size of a packed polynomial.
-	PolySize int = 384
+	PolySize = params.PolySize
 
 	// PlaintextSize is the size of the plaintext
-	PlaintextSize = 32
+	PlaintextSize = params.PlaintextSize
 )