--- a/pke/kyber/internal/common/stubs_amd64.go
+++ b/pke/kyber/internal/common/stubs_amd64.go
@@ -9,3 +9,6 @@ func addAVX2(p *[256]int16, a *[256]int16, b *[256]int16)
 
 //go:noescape
 func subAVX2(p *[256]int16, a *[256]int16, b *[256]int16)
+
+//go:noescape
+func nttAVX2(p *[256]int16)