… pke/kyber/internal/common/amd64.s
… pke/kyber/internal/common/asm/go.sum
… pke/kyber/internal/common/poly_test.go