Speeds up NTT by 20x.  My approach is similar to that of the reference AVX2 implementation.  It might be the same, but it's hard to say: I had to rediscover big parts as their documentation is rather lacking.  I tried to make this implementation easier to read, although I must say that this NTT is more challenging than that of Dilithium.

[<img width="1080" alt="Ek895X9XYAYR0Hv" src="https://user-images.githubusercontent.com/9975/97027720-6e336000-155b-11eb-9774-6ca8d958715b.png">](https://westerbaan.name/~bas/images/kybavx2.svg.gz)

See #158