@bwesterb 2020-10-24
> are there any benchmarks for kyber to see the impact on kem operations?

I added benchmarks and also ran them on master:

This PR:

```
BenchmarkGenerateKey/Kyber512-4         	   35932	     32349 ns/op
BenchmarkGenerateKey/Kyber768-4         	   19324	     60656 ns/op
BenchmarkGenerateKey/Kyber1024-4        	   12500	     94076 ns/op
BenchmarkEncapsulate/Kyber512-4         	   36661	     32673 ns/op
BenchmarkEncapsulate/Kyber768-4         	   25603	     46045 ns/op
BenchmarkEncapsulate/Kyber1024-4        	   19406	     61983 ns/op
BenchmarkDecapsulate/Kyber512-4         	   30687	     39616 ns/op
BenchmarkDecapsulate/Kyber768-4         	   20198	     54756 ns/op
BenchmarkDecapsulate/Kyber1024-4        	   16562	     71975 ns/op
```

On master (old):

```
BenchmarkGenerateKey/Kyber512-4         	   27952	     42393 ns/op
BenchmarkGenerateKey/Kyber768-4         	   15742	     76090 ns/op
BenchmarkGenerateKey/Kyber1024-4        	   10000	    127672 ns/op
BenchmarkEncapsulate/Kyber512-4         	   28323	     42978 ns/op
BenchmarkEncapsulate/Kyber768-4         	   19965	     64035 ns/op
BenchmarkEncapsulate/Kyber1024-4        	   15351	     77215 ns/op
BenchmarkDecapsulate/Kyber512-4         	   19413	     55272 ns/op
BenchmarkDecapsulate/Kyber768-4         	   15804	     72707 ns/op
BenchmarkDecapsulate/Kyber1024-4        	   12530	     95090 ns/op
```

This is just the forward NTT.  With InvNTT, MulHat and others AVX2 optimised, this will be really quite fast.