@armfazh 2020-10-26
This is the report using benchcmp (https://godoc.org/golang.org/x/tools/cmd/benchcmp)

```md
benchmark                            old ns/op     new ns/op     delta
BenchmarkGenerateKey/Kyber512-4      42393         32349         -23.69%
BenchmarkGenerateKey/Kyber768-4      76090         60656         -20.28%
BenchmarkGenerateKey/Kyber1024-4     127672        94076         -26.31%
BenchmarkEncapsulate/Kyber512-4      42978         32673         -23.98%
BenchmarkEncapsulate/Kyber768-4      64035         46045         -28.09%
BenchmarkEncapsulate/Kyber1024-4     77215         61983         -19.73%
BenchmarkDecapsulate/Kyber512-4      55272         39616         -28.33%
BenchmarkDecapsulate/Kyber768-4      72707         54756         -24.69%
BenchmarkDecapsulate/Kyber1024-4     95090         71975         -24.31%

```